feat(cinder-understack): create minimal NVMe storage library wrapper
Wrap the upstream NetAppNVMeStorageLibrary with a minimal wrapper which
allows us to reuse the upstream class on a per-SVM basis from our
multi-SVM supporting driver. The __init__() and the do_setup()
specifically had calls in them which resulted in our code not working
due to the configuration override we must do.