chore(cinder-understack): add functions to track SVM for a pool
We need to be able to track the SVM with a pool. Unfortunately flexvols
which are 1:1 with pool names for a backend are unique in a SVM but they
can be repeated between SVMs. So we must prefix the pool name with our
SVM name and then strip it off when we need just the pool name.