sync_keystone: Stop managing OUTSIDE networks for tenants
We needed to create this network on behalf of the tenant because it
required admin privs to make an "external" network.   We can now achieve
the same effect using a normal network that a tenant can create for
themselves.

The user will create a network, subnet and router.  This has the same
end result, and therefore we are no longer going to support the OUTSIDE
network.

The intention is to proactively delete all existing OUTSIDE networks, so I
have gone ahead and removed the cleanup code from here.