project.created sensor - ignore untagged projects
Prior to this commit, we relied on a "create SVM" workflow checking if
the newly minted project has UNDERSTACK_SVM tags. If it didn't, the
workflow terminated with non-zero exit code indicating failure and this
prevented the Ansible workflow from executing.

But creating a project without the tags should not be considered an
exception or erroneous situation, because it may create confusion,
unnecessary alerts and so on. Instead of that, we simply read the output
of the previous workflow to see if Ansible should proceed or not.