feat(argo-workflows): don't run as namespaced
Make the Argo Server, the UI and API component, and the Workflow
Controller, the actual workflow engine, not run as namespaced so the UI
and API can see and manage Argo Events. This is because Argo Events uses
the same Argo Server for its UI and we run the EventSources and Sensors
in a different namespace. Changed the Workflow Controller as well
because it was easier and cleaner to do so and allowed us to use the
upstream manifests more cleanly. To achieve the same protection and
separation, switch to the ClusterRole and just use a RoleBinding to bind
it to specific namespaces. This makes using the upstream manifests a
little more cleaner and they already have separate roles for what we
called the runtime so we can delete those.