feat(argo-workflows): update and document SSO and Ingress
We won't be able to make this dynamic and do substitution because that
goes against the nature of how kustomize likes to have everything
declarative. The dns_zone value was a hack and it doesn't work outside
of an AIO setup. Document for the user how they can get things setup and
drop applying the SSO and Ingress settings, which aren't necessary if
you don't want to hit the UI or the API. Renamed the directory to make
it more clear what part this is and be consistent with the app name.