react on svm_created, not svm_enabled
- svm_enabled is a variable/output that tells if particular project is
  configured to have an SVM created in NetApp
- svm_created is a variable that records if the SVM was created in a
  particular workflow run (not that it exists or that is expected to
  exist).

This distinction allows us to gracefully react to a situation where an
end-user updates a project by adding additional tag that is not NetApp
related. Prior to this commit, this would trigger a workflow that
checks if the SVM already exists and that first part of the workflow
(oslo_event handler specifically) would terminate gracefully without a
problem. But in the next step, the ansible-to-nautobot workflow would
still get triggered even though it should not.