chore: drop dns_zone from site configuration
The only place this was used was for Argo Workflows but we need the Dex
URL and we need the Ingress URL for the service so having this as one
value would have never worked. To actually do this you need to use the
documentation configuration. This leaves dns_zone in place, because
that's still used in the ArgoCD deployment.