feat(keystone): add OAuth 2.0 authentication options
The device authorization and the client credentials flows are OAuth 2.0
flows instead of OIDC so we need to be able to validate the token
submitted against the endpoint so we need to also read the metadata for
the OAuth 2.0 paths of mod_auth_openidc. Add another authentication
endpoint into apache for Keystone so that we can route the
authentication request to the correct connector inside of Dex to
successfully complete the authentication.