fix(keystone): avoid resetting keys on each deploy
Due to the way OpenStack Helm sets the helm.sh/hook annotation on the
secrets always it causes a removal and re-apply of the secrets but
dropping all the generated data every time the keystone chart is
deployed. We do not want this behavior so disable OpenStack Helm
creating this secret for us and instead switch to just loading the
secret ourselves. ref:PUC-1118