chore(openstack): create schema and document configuration
Created a schema for the current chart and validated samples we have
against it. Updated a few fields from JSON to YAML to make it consistent
and let my validation test pass. Symlink the schema so that it is able
to be fetched. Added documentation around how to configure this part of
a deployment.