Remove Nautobot integration from ml2 mechanism driver
The driver was performing synchronous updates to keep Nautobot
up-to-date with changes to networks, prefixes, etc.  This approach has a
number of drawbacks.

We are no longer treating Nautobot as the authoritative source for IP
space, network names, etc.   Whenever a user is prevented from doing
something because Nautobot said "no", it is always seen as a problem or
bug in our system.  The expected behaviour is that all valid openstack
operations should succeed.

Here we remove all Nautobot integration from the driver, with the
intention of updating Nautobot asynchronously using a new mechanism.