feat(openstack): share service acct credentials via ESO
Utilize the External Secrets Operator to define the OpenStack service
account credentials that OpenStack Helm needs to work with Keystone
service accounts for the various OpenStack services. This allows one
cluster to run Keystone and another to run the OpenStack services and to
utilize the ESO operator to keep the credentials in sync between the
two.

Co-authored-by: Marek Skrobacki <marek.skrobacki@rackspace.co.uk>