fix: handle keys with dot in the name
Argo Events does not support accessing the event data if the key
includes the dot because there is no known way to escape that dot and
the square bracket notation cannot be used in dataKey or contextKey.

This workaround simply copies the value of the ironic_object.data into
body.ironic_object. We could theoretically delete the old copy, but I'm
keeping it around in case the original value needs to be consumed.