fix(ironic): workaround hardcoded --config-file flag in scripts
This should have been part of ad141406da. It's commit message is below.

The --config-file flag is hardcoded in a number of scripts which results
in preference to that file over the directories which is not the
behavior we want since we want to override. However since the OpenStack
Helm charts provide no way to change this behavior we need to replace
the entire configmap-bin. This one is taken from two environments which
were identical and then had the --config-dir flag added so that it
continues to work.