feat(openstack): mount the keystone secrets config snippets
This was missed from c589b4. It's commit message is below.

The change to mount these has not landed in all upstream charts and
we've also not updated to the charts that have this change across the
board. So for now include it in all of our configs until this change can
make its way upstream.