fix: flavorsync service acct role
While the flavors themselves are not scoped to a specific project, the
project scoped tokens by default are issued for the project that the
user is in. As a result, if we are creating 'flavorsync' service account
in 'service' project, we should assign the role there too instead of the
default.