chore(neutron-understack): switch oslo_config overrides to a fixture
Rather than patching the global scoped oslo_config, switch to using a
fixture which resets between tests to ensure that things are cleaned up
appropriately. This will be necessary when we no longer initialize
oslo_config in the global/import scope of the module. While moving this
it pointed out an issue with one of the tests missing a mock which
attempted to hit the DB, which now couldn't be initialized because we
didn't setup our configs because the data had the wrong types which is
being fixed in this as well.