fix(nautobot): multiple corrections to nautobot sync job
The nautobot sync job never had a TTL set to clean up after it ran so
now we clean up 5 minutes after it runs. It should also have a safety
that it should not run for longer than 20 minutes so this defines that
limit. Added a few more helm annotations for cleanup of the job and for
being able to lookup the jobs by labels.

Ensure the container runs securely by specifying a permission limited
security context.

Lastly fix the device-types role by mounting the data that the role
needs into the container.