feat(hardware categorization): remove CPU core count from matching rules
Unfortunately the IPA image reports the number of CPU threads and not
the number of CPU cores without giving us the ability to calculate the
cores because there's no value to say that hyperthreading is enabled or
not. The model of the CPU should be detailed enough for us to be happy
that this is correct so we'll just drop CPU core count.