feat(ironic): add pci-devices and disable physical-network inspect hook
Removed the physical-network inspection hook. The physical-network hook
requires specific configuration to map the physical-network based on
the IP addresses that the provisioning network is and this likely won't
be the best way to determine the physical network. Add pci-devices hook
to load the PCI device information into the inventory.