chore: update OpenStack container builds to shrink them
Switched to using more of a two part build in the containers to install
everything in the build layer and then copy it into the runtime
container to shrink them further. Switch to using uv exclusively to make
it consistent with the build tooling we use. Add the OpenStack
constraint file to ensure we do not upgrade to too new of a version when
installing additional dependencies.