feat: Add an ArgoCD Application chart to replace AppSets
The AppSets have proven to have quite a bit of limitations and attempts
to make the kustomize interface take references like the helm interface
can from different repos or branches has not merged upstream. It's
harder to see what is being generated or rendered by the AppSet as well.
ArgoCD's UI doesn't show AppSets or what they are doing but shows
Applications in detail. So switch to utilizing Helm to generate the
Applications. We have the added benefit that the Applications can be of
different base templates in the same ArgoCD for targeting different
clusters as well. With AppSets we have the same base template that is
used for all the clusters so they need to stay relatively at the same
version.