Configure ironic Inspection hooks to manage port and node attributes
After out-of-band inspection we run port-bios-name.

The interface names returned by the BMC are generally more meaningful to
the data center than the linux names we get in the agent inspection data

After Agent Inspection we run update-baremetal-port.

This consumes the LLDP information, so it must run after agent
inspection.  Out-of-band inspection doesn't give us the full LLDP data.

The default ironic local-link-connection is removed because the new hook
updates the same fields.

We add "validate-interfaces" because that populates data we consume.