Set local_link_connection for all ports we know about
We document all available connections in the local_link_connection
field.  However we ONLY ever populate the physical_network field for
VLAN Groups whose names end with "-network".

The physical_network field is used to determine which NICs are eligible
for use by neutron, so don't populate it unless we are happy for the NIC
will work for tenant VNIs.